parse components as fragments and wrap them in `<template>` tags
The use case that lead to this was a component for a table row.  a
`<tr>` is not allowed inside a `<body>` or `<html>`, nor can it be the
start of a fragment.  The solution is to tell Nokogiri to pretend it's
wrapped in a `<template>` element, which can contain anything.