don't require traceparent
Apparently, if this code is run outside of an existing span, the
traceparent "carrier" will be an empty hash.  This would happen when
running a test of a page or component that uses that uses component.

So, if there is no traceparent, we'll still render the component, but
with some diagnostic information.