rework MCP to be a bit cleaner + fix puma issues.
Puma does not like when you make a thread before workers have booted.
SemanticLogger does this, however it can be set up to not cause a
problem.  To avoid a warning from Puma, we set the thread local variable
that puma uses to know it's OK to not warn about.