move gems needed only for dev out of runtime
I hoped this would reduce the need to explicitly require these gems in
the developer's app, but it ends up requiring them in all groups.
The most acute problem this causes is that Faker's massive list of YAML
files is loaded into I18n and it kills app performance.

A follow-on consequence of this is that seed data cannot be auto-loaded,
since this would transitively require Faker.