improve behavior for checkboxes.
With one checkbox, the behavior is OK - either  "true" is provided or
nothing is, which is interpreted as false.

But for an array of checkboxes, this doesn't work. This solution ensures
that the checkbox values are indexes into the array and then transforms
those indexes into an array of true/nil values to match the other form
elements.