re-order to avoid trying to use OTel too early + improve shutdown
The various `bin/db` commands use exceptions for control flow in order
to allow bootstrapping to work but not blow up if the DB is not a good
state.

This is probably not a great pattern, however it is currently what it
is.  Thus, we want OTel set up first, so if boot gets interrupted by
postgres issues, we don't try to use an un-configured OTel.